requirements.txt
fastapi 
uvicorn  
python-dotenv
together
gunicorn
sqlalchemy 
psycopg2
databases 
asyncpg
faker